cards-rendering(js)
adde cards-rendering.js file with movie card rendering
deletid function renderMoviesCards from catalog.js
